fix(widgetbook): split dialogue overlay stories into catalog_part4.dart for repo.part_unit_size (Refs #2753)
PR #2759's earlier slices pushed catalog_part3.dart to 1049 physical lines,
exceeding the 1000-line cap enforced by repo.part_unit_size. Move the
new dialogue-overlay helpers, fixtures, and Widgetbook directories into a
new catalog_part4.dart and register the part directive in catalog.dart:

- _InlineYarnAssetBundle
- _CtDialogueViewStoryHarness + ctDialogueViewDirectories
- _kCtDialogueViewStorySource / _kGameStartIntroOverlayStorySource
- gameStartIntroOverlayDirectories
- _overtureStoryGame + overtureDialogueOverlayDirectories
- _callToArmsStoryGame + callToArmsDialogueOverlayDirectories

Verification:
- dart run tool/check_part_unit_size.dart -> no violations.
- CT_REPO_LINT_INCLUDE_APP=true dart run tool/ct_repo_lint.dart -> all 47 rule(s) passed.
- cd app && flutter analyze lib/widgetbook/ test/dialogue_overlays_specs_test.dart -> No issues found.
- cd app && flutter test test/dialogue_overlays_specs_test.dart -> 14/14 pass.
